feat: add domain service for code generation templates
