+++ spring-ddd-domain/src/main/java/com/springddd/domain/gen/GenerateDomainService.java
@@ -1,0 +1,8 @@
+package com.springddd.domain.gen;
+
+import reactor.core.publisher.Mono;
+
+public interface GenerateDomainService {
+
+    Mono<Void> generate(String tableName);
+}
